+++ test/test_combinatorics.js
@@ -80,6 +80,24 @@ QUnit.test('Next subset computation', function(assert) {
 });
 
 
+
+QUnit.test('Next k-subset computation', function(assert) {    
+  // Reference: Nijenhuis, A., & Wilf, H. S. (1978). Combinatorial algorithms for computers and calculators. 2d ed. New York: Academic Press.
+  // Test with the static data examples of section 3
+  {
+	  var expectedValues = [[true,[1,2,3]], [true,[1,2,4]], [true,[1,2,5]], [true,[1,3,4]], [true,[1,3,5]], 
+	                        [true,[1,4,5]], [true,[2,3,4]],[true,[2,3,5]], [true,[2,4,5]], [false,[3,4,5]]];
+	  
+	  var nextKSubsetIterator = new PortfolioAllocation.kSubsetsIterator_(5, 3);
+	  for (var i = 0; i < expectedValues.length; ++i) {
+		var nextKSubset = nextKSubsetIterator.next();
+		assert.deepEqual(nextKSubset, expectedValues[i], 'Next k-subset - Test 1 #' + i);
+	  }  
+  }
+
+});
+
+
 QUnit.test('Next random k-subset computation', function(assert) {    
   // Test with random data
   {